--- a/cmd/cayley/cayley.go
+++ b/cmd/cayley/cayley.go
@@ -54,6 +54,7 @@ import (
 	_ "github.com/cayleygraph/cayley/writer"
 
 	// Load supported query languages
+	_ "github.com/cayleygraph/cayley/query/gizmo"
 	_ "github.com/cayleygraph/cayley/query/gremlin"
 	_ "github.com/cayleygraph/cayley/query/mql"
 	_ "github.com/cayleygraph/cayley/query/sexp"