… glide.yaml
… integration/integration_test.go
… query/gizmo/traversals.go
… templates/sidebar.tmpl