Switch JS VM from Otto to Goja and name this query language Gizmo.

Benchmarks (with https://github.com/dop251/goja/pull/12 merged):
```
benchmark                                     old ns/op     new ns/op     delta
BenchmarkNamePredicate-8                      169447        133728        -21.08%
BenchmarkLargeSetsNoIntersection-8            9474013       8246870       -12.95%
BenchmarkVeryLargeSetsSmallIntersection-8     23505905      20423973      -13.11%
BenchmarkHelplessContainsChecker-8            34720988      31287961      -9.89%
BenchmarkHelplessNotContainsFilms-8           35750687      32215110      -9.89%
BenchmarkHelplessNotContainsActors-8          37780076      33983268      -10.05%
BenchmarkNetAndSpeed-8                        1494604       1103341       -26.18%
BenchmarkKeanuAndNet-8                        814971        697612        -14.40%
BenchmarkKeanuAndSpeed-8                      1226941       923928        -24.70%
BenchmarkKeanuOther-8                         5932712       5231494       -11.82%
BenchmarkKeanuBullockOther-8                  10137872      8801214       -13.18%
BenchmarkSaveBogartPerformances-8             352812        345900        -1.96%

benchmark                                     old allocs     new allocs     delta
BenchmarkNamePredicate-8                      317            483            +52.37%
BenchmarkLargeSetsNoIntersection-8            39859          40054          +0.49%
BenchmarkVeryLargeSetsSmallIntersection-8     96119          93806          -2.41%
BenchmarkHelplessContainsChecker-8            142001         142216         +0.15%
BenchmarkHelplessNotContainsFilms-8           145570         145835         +0.18%
BenchmarkHelplessNotContainsActors-8          151542         151795         +0.17%
BenchmarkNetAndSpeed-8                        6041           6328           +4.75%
BenchmarkKeanuAndNet-8                        3803           4088           +7.49%
BenchmarkKeanuAndSpeed-8                      4840           5127           +5.93%
BenchmarkKeanuOther-8                         23492          23802          +1.32%
BenchmarkKeanuBullockOther-8                  38644          38972          +0.85%
BenchmarkSaveBogartPerformances-8             1306           1496           +14.55%

benchmark                                     old bytes     new bytes     delta
BenchmarkNamePredicate-8                      24725         32057         +29.65%
BenchmarkLargeSetsNoIntersection-8            3578938       3598218       +0.54%
BenchmarkVeryLargeSetsSmallIntersection-8     8952836       8819102       -1.49%
BenchmarkHelplessContainsChecker-8            14898533      14916653      +0.12%
BenchmarkHelplessNotContainsFilms-8           15124003      15150157      +0.17%
BenchmarkHelplessNotContainsActors-8          15352874      15376615      +0.15%
BenchmarkNetAndSpeed-8                        453997        492152        +8.40%
BenchmarkKeanuAndNet-8                        251676        289547        +15.05%
BenchmarkKeanuAndSpeed-8                      350401        388555        +10.89%
BenchmarkKeanuOther-8                         2122510       2167150       +2.10%
BenchmarkKeanuBullockOther-8                  3518228       3562027       +1.24%
BenchmarkSaveBogartPerformances-8             103643        114399        +10.38%
```
Still need to figure out why Goja allocates that much.